Int array to double array

+++ include/structs.h
@@ -174,6 +174,6 @@ void MakeCliqueHT(Link, HashTable*, int );
 void CliqueConcat(Pair*, Pair*, int,int );
 void ChangeUnrelated(Link, HashTable*, int );
 void printUnrelated(Link ,FILE* ,char*,int );
-void CreateArray( Link , HashTable* , int, int** );
+void CreateArray( Link , HashTable* , int, double** );
