+++ include/text_support.h
@@ -11,10 +11,10 @@ typedef struct{
 Item* parse(char*);
 void read_csv(HashTable* ,char*, int);
 void textCleaning(char*);
-void tokenize(char* , HashTable* , HashTable*  ,int , int ** , int* );
+void tokenize(char* , HashTable* , HashTable*  ,int , double** , int* );
 void read_stopwords(HashTable* ,char* ,int );
 void CreateDictionary(Stats* , HashTable* ,HashTable* , int);
-void UpdateArray(Stats*  , HashTable*  , int , int** );
-double** Bow_To_Tfidf(int ** , int , int );
+void UpdateArray(Stats*  , HashTable*  , int , double** );
+double** Bow_To_Tfidf(double** , int , int );
 void CutOffDictionary(HashTable*,int , int);
